+++ src/errorHandlers/index.ts
@@ -1,5 +1,6 @@
 export * from './handle400WithValidationsFunc';
 export * from './handle404Func';
 export * from './handle4xxFunc';
+export * from './handleHttp404Error';
 export * from './handleHttpClientErrorResponse';
 export * from './handleHttpClientErrors';